# openapi/user-storage-bff.yaml
openapi: 3.1.0
info:
  title: KYC Backend BFF Flow API
  version: 0.2.2
  description: >
    BFF-facing v2 flow runtime API. Only flow sessions, flows, steps, user reads, and upload endpoints remain on this surface.
servers:
  - url: https://kyc-internal.example.local
...
security:
  - KcSignature: [ ]
paths:
...
  /flow/sessions:
    get:
      tags: [ sessions ]
      summary: List flow sessions for the caller
      operationId: listFlowSessions
      responses:
        "200":
          description: Sessions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FlowSession"
    post:
      tags: [ sessions ]
      summary: Create a top-level flow session
      operationId: createFlowSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFlowSessionRequest"
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlowSession"
  /flow/sessions/{sessionId}:
    get:
      tags: [ sessions ]
      summary: Get session with nested flows
      operationId: getFlowSession
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Session detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlowSessionDetail"
...
  /flow/flows/{flowId}:
    get:
      tags: [ flows ]
      summary: Get flow with its steps
      operationId: getFlow
      parameters:
        - $ref: "#/components/parameters/FlowId"
      responses:
        "200":
          description: Flow detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlowDetail"
  /flow/flows/{flowId}/steps:
    get:
      tags: [ steps ]
      summary: List steps for a flow
      operationId: listFlowSteps
      parameters:
        - $ref: "#/components/parameters/FlowId"
      responses:
        "200":
          description: Flow steps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FlowStep"
  /flow/steps/{stepId}:
    get:
      tags: [ steps ]
      summary: Get flow step details
      operationId: getFlowStep
      parameters:
        - $ref: "#/components/parameters/StepId"
      responses:
        "200":
          description: Step detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlowStep"
    post:
      tags: [ steps ]
      summary: Submit user input for a waiting step
      operationId: submitFlowStep
      parameters:
        - $ref: "#/components/parameters/StepId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitFlowStepRequest"
      responses:
        "200":
          description: Step updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlowStep"
components:
  parameters:
...
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
        minLength: 8
    FlowId:
      name: flowId
      in: path
      required: true
      schema:
        type: string
        minLength: 8
    StepId:
      name: stepId
      in: path
      required: true
      schema:
        type: string
        minLength: 8
  schemas:
...
    CreateFlowSessionRequest:
      type: object
      additionalProperties: false
      required: [ sessionType ]
      properties:
        sessionType:
          type: string
        humanId:
          type: string
        context:
          type: object
          additionalProperties: true
...
    SubmitFlowStepRequest:
      type: object
      additionalProperties: false
      properties:
        input:
          type: object
          additionalProperties: true
    FlowSession:
      type: object
      additionalProperties: false
      required: [ id, humanId, sessionType, status, context, createdAt, updatedAt ]
      properties:
        id: { type: string }
        humanId: { type: string }
        sessionType: { type: string }
        status: { type: string }
        userId:
          type: string
          nullable: true
        context:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    FlowInstance:
      type: object
      additionalProperties: false
      required: [ id, humanId, sessionId, flowType, status, stepIds, context, createdAt, updatedAt ]
      properties:
        id: { type: string }
        humanId: { type: string }
        sessionId: { type: string }
        flowType: { type: string }
        status: { type: string }
        currentStep:
          type: string
          nullable: true
        stepIds:
          type: array
          items:
            type: string
        context:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    FlowStep:
      type: object
      additionalProperties: false
      required: [ id, humanId, flowId, stepType, actor, status, attemptNo, createdAt, updatedAt ]
      properties:
        id: { type: string }
        humanId: { type: string }
        flowId: { type: string }
        stepType: { type: string }
        actor: { type: string }
        status: { type: string }
        attemptNo: { type: integer }
        input:
          type: object
          additionalProperties: true
          nullable: true
        output:
          type: object
          additionalProperties: true
          nullable: true
        error:
          type: object
          additionalProperties: true
          nullable: true
        nextRetryAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
          nullable: true
    FlowSessionDetail:
      type: object
      additionalProperties: false
      required: [ session, flows ]
      properties:
        session:
          $ref: "#/components/schemas/FlowSession"
        flows:
          type: array
          items:
            $ref: "#/components/schemas/FlowInstance"
    FlowDetail:
      type: object
      additionalProperties: false
      required: [ flow, steps ]
      properties:
        flow:
          $ref: "#/components/schemas/FlowInstance"
        steps:
          type: array
          items:
            $ref: "#/components/schemas/FlowStep"
  securitySchemes:
    KcSignature:
      type: apiKey
      in: header
      name: x-kc-signature